+++ ReleaseNotes/release_notes.md
@@ -9,6 +9,7 @@
  1. Add `Ostap::Math::identric_mean`
  1. Add `Ostap::Math::logarithmic_mean`
  1. Add `Ostap::Math::stolarsky_mean` 
+ 1. Extend Fermi-Dirac,Bose-Einstein and Debye functions, add them to `math_ve` 
  
 
 ## Bug fixes